function-app/pyproject.toml
[project]
name = "function-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.24.0",
]
